src/main/java/hackerrank/algorithm/medium/ExtraLongFactorials.java
package hackerrank.algorithm.medium;

import java.math.BigInteger;
import java.util.Scanner;


public class ExtraLongFactorials {

    // Complete the extraLongFactorials function below.
    static void extraLongFactorials(int n) {
        System.out.println(factorial(n));
    }

    static BigInteger factorial(int n) {
        if (n == 1) {
            return BigInteger.valueOf(1);
        } else {
            return BigInteger.valueOf(n).multiply(factorial(n - 1));
        }
    }

    private static final Scanner scanner = new Scanner(System.in);

    public static void main(String[] args) {
        int n = scanner.nextInt();
        scanner.skip("(\r\n|[\n\r\u2028\u2029\u0085])?");

        extraLongFactorials(n);

        scanner.close();
    }
}

